+++ mpire/utils.py
@@ -16,7 +16,7 @@ except ImportError:
     np = None
     NUMPY_INSTALLED = False
 
-from mpire.context import RUNNING_MACOS, RUNNING_WINDOWS
+from mpire.context import RUNNING_MACOS, RUNNING_WINDOWS, mp_dill
 
 # Needed for setting CPU affinity
 if RUNNING_WINDOWS:
@@ -249,14 +249,28 @@ class TimeIt:
             heapq.heappushpop(self.max_time_array,
                               (duration, self.format_args_func() if self.format_args_func is not None else None))
 
+
+def create_sync_manager(use_dill: bool) -> SyncManager:
+    """
+    Create a SyncManager instance
+
+    :param use_dill: Whether dill is used as serialization library
+    :return: SyncManager instance
+    """
+    authkey = os.urandom(24)
+    return mp_dill.managers.SyncManager(authkey=authkey) if use_dill else SyncManager(authkey=authkey)
+
     
 class NonPickledSyncManager:
     """ SyncManager wrapper that won't be pickled """
     
-    def __init__(self) -> None:
-        self.manager = SyncManager()
+    def __init__(self, use_dill: bool) -> None:
+        """
+        :param use_dill: Whether dill is used as serialization library
+        """
+        self.manager = create_sync_manager(use_dill)
         
-    def __getattr__(self, item):
+    def __getattr__(self, item: str):
         return getattr(self.manager, item)
 
     def __getstate__(self) -> dict: